+++ android/app/build.gradle
@@ -37,7 +37,7 @@ react {
     //   The name of the generated asset file containing your JS bundle
     // bundleAssetName = "MyApplication.android.bundle"
     //
-    //   The entry file for bundle generation. Default is 'index.android.js' or 'index.ts'
+    //   The entry file for bundle generation. Default is 'index.android.js' or 'store.ts'
     // entryFile = file("../js/MyApplication.android.js")
     //
     //   A list of extra flags to pass to the 'bundle' commands.
