+++ src/local-store/FavoriteStore.ts
@@ -1,7 +1,7 @@
 import { create } from 'zustand';
-import { MovieDetail } from '../../types/MovieDetail.ts';
+import { MovieDetail } from '../types/MovieDetail.ts';
 import { persist, createJSONStorage } from 'zustand/middleware';
-import { TvSeriesDetail } from '../../types/TvSeriesDetail.ts';
+import { TvSeriesDetail } from '../types/TvSeriesDetail.ts';
 import { zustandMMKVStorage } from './mmkv.ts';
 
 interface FavoriteMoviesTvSeriesStore {
